requirements.txt
telethon>=1.36.0
pandas>=2.0.0
vaderSentiment>=3.3.2
matplotlib>=3.7.0
seaborn>=0.13.0
wordcloud>=1.9.0
Jinja2>=3.1.0
python-dotenv>=1.0.0
rich>=13.0.0
